start to retire getTwilioAuth
